Update asmjit to latest version and fix API compatibility
- Upgrade asmjit library with new support/ directory structure
- Fix MemoryManipulation.cpp: x86::Reg -> asmjit::Reg, x86::Xmm -> x86::Vec,
  cpuFeatures() -> cpu_features(), setLogger() -> set_logger(),
  Error type now requires explicit comparison
- Fix Data.h: Add using declaration to inherit variant constructors
- Fix AI.cpp: Cast OutcomeType enum to int for LuaVariant
- Update vcxproj: Replace deleted zone*.cpp with new support/*.cpp files,
  add capstone.lib to linker dependencies

+++ AI.cpp
@@ -17,7 +17,7 @@ namespace AI::CongressSupport {
 
 		if (DoesProcessorExist(name)) {
 			auto variantMap = LuaVariantMap();
-			variantMap.emplace("OutcomeType", LuaVariant(outcomeType));
+			variantMap.emplace("OutcomeType", LuaVariant(static_cast<int>(outcomeType)));
 			variantMap.emplace("PlayerId", LuaVariant(*(int*)((uintptr_t)player + 0xd8)));
 			variantMap.emplace(decisionKey, LuaVariant(*(int*)((uintptr_t)modifierAnalysis + fieldOffset)));
